+++ JPEGImageCompression/reducible_colors.py
@@ -1,0 +1,9 @@
+REDUCIBLE_PURPLE_DARKER = "#3B0893"
+REDUCIBLE_BLUE = "#650FFA"
+REDUCIBLE_PURPLE = "#8c4dfb"
+REDUCIBLE_VIOLET = "#d7b5fe"
+REDUCIBLE_YELLOW = "#ffff5c"
+REDUCIBLE_YELLOW_DARKER = "#7F7F2D"
+REDUCIBLE_GREEN_LIGHTER = "#00cc70"
+REDUCIBLE_GREEN = "#008f4f"
+REDUCIBLE_GREEN_DARKER = "#004F2C"
