+++ JPEGImageCompression/scenes.py
@@ -1,9 +1,14 @@
+"""
+File to define all the different scenes our video will be composed of.
+"""
+
 from math import sqrt
 from manim import *
 import cv2
-from pprint import pprint
-from scipy import fftpack
-from typing import Iterable, List
+
+from functions import *
+from classes import *
+from reducible_colors import *
 
 config["assets_dir"] = "assets"
 np.random.seed(23)
@@ -14,196 +19,6 @@ If you run with caching, since there are some scenes that change pixel arrays,
 there might be some unexpected behavior
 E.g manim -pql JPEGImageCompression/image_compression.py --disable_caching
 """
-
-REDUCIBLE_PURPLE_DARKER = "#3B0893"
-REDUCIBLE_BLUE = "#650FFA"
-REDUCIBLE_PURPLE = "#8c4dfb"
-REDUCIBLE_VIOLET = "#d7b5fe"
-REDUCIBLE_YELLOW = "#ffff5c"
-REDUCIBLE_YELLOW_DARKER = "#7F7F2D"
-REDUCIBLE_GREEN_LIGHTER = "#00cc70"
-REDUCIBLE_GREEN = "#008f4f"
-REDUCIBLE_GREEN_DARKER = "#004F2C"
-
-
-class ReducibleBarChart(BarChart):
-    """
-    Redefinition of the BarChart class to add font personalization
-    """
-
-    def __init__(
-        self,
-        values: Iterable[float],
-        height: float = 4,
-        width: float = 6,
-        n_ticks: int = 4,
-        tick_width: float = 0.2,
-        chart_font: str = "SF Mono",
-        label_y_axis: bool = True,
-        y_axis_label_height: float = 0.25,
-        max_value: float = 1,
-        bar_colors=...,
-        bar_fill_opacity: float = 0.8,
-        bar_stroke_width: float = 3,
-        bar_names: List[str] = ...,
-        bar_label_scale_val: float = 0.75,
-        **kwargs,
-    ):
-        self.chart_font = chart_font
-
-        super().__init__(
-            values,
-            height=height,
-            width=width,
-            n_ticks=n_ticks,
-            tick_width=tick_width,
-            label_y_axis=label_y_axis,
-            y_axis_label_height=y_axis_label_height,
-            max_value=max_value,
-            bar_colors=bar_colors,
-            bar_fill_opacity=bar_fill_opacity,
-            bar_stroke_width=bar_stroke_width,
-            bar_names=bar_names,
-            bar_label_scale_val=bar_label_scale_val,
-            **kwargs,
-        )
-
-    def add_axes(self):
-        x_axis = Line(self.tick_width * LEFT / 2, self.total_bar_width * RIGHT)
-        y_axis = Line(ORIGIN, self.total_bar_height * UP)
-        ticks = VGroup()
-        heights = np.linspace(0, self.total_bar_height, self.n_ticks + 1)
-        values = np.linspace(0, self.max_value, self.n_ticks + 1)
-        for y, _value in zip(heights, values):
-            tick = Line(LEFT, RIGHT)
-            tick.width = self.tick_width
-            tick.move_to(y * UP)
-            ticks.add(tick)
-        y_axis.add(ticks)
-
-        self.add(x_axis, y_axis)
-        self.x_axis, self.y_axis = x_axis, y_axis
-
-        if self.label_y_axis:
-            labels = VGroup()
-            for tick, value in zip(ticks, values):
-                label = Text(str(np.round(value, 2)), font=self.chart_font)
-                label.height = self.y_axis_label_height
-                label.next_to(tick, LEFT, SMALL_BUFF)
-                labels.add(label)
-            self.y_axis_labels = labels
-            self.add(labels)
-
-    def add_bars(self, values):
-        buff = float(self.total_bar_width) / (2 * len(values) + 1)
-        bars = VGroup()
-        for i, value in enumerate(values):
-            bar = Rectangle(
-                height=(value / self.max_value) * self.total_bar_height,
-                width=buff,
-                stroke_width=self.bar_stroke_width,
-                fill_opacity=self.bar_fill_opacity,
-            )
-            bar.move_to((2 * i + 1) * buff * RIGHT, DOWN + LEFT)
-            bars.add(bar)
-        bars.set_color_by_gradient(*self.bar_colors)
-
-        bar_labels = VGroup()
-        for bar, name in zip(bars, self.bar_names):
-            label = Text(str(name), font="SF Mono")
-            label.scale(self.bar_label_scale_val)
-            label.next_to(bar, DOWN, SMALL_BUFF)
-            bar_labels.add(label)
-
-        self.add(bars, bar_labels)
-        self.bars = bars
-        self.bar_labels = bar_labels
-
-
-class Module(VGroup):
-    def __init__(
-        self,
-        text,
-        fill_color=REDUCIBLE_PURPLE_DARKER,
-        stroke_color=REDUCIBLE_VIOLET,
-        stroke_width=5,
-        text_scale=0.9,
-        text_position=ORIGIN,
-        text_weight=NORMAL,
-        width=4,
-        height=2,
-        **kwargs,
-    ):
-
-        self.rect = (
-            RoundedRectangle(
-                corner_radius=0.1, fill_color=fill_color, width=width, height=height
-            )
-            .set_opacity(1)
-            .set_stroke(stroke_color, width=stroke_width)
-        )
-
-        self.text = Text(str(text), weight=text_weight, font="CMU Serif").scale(
-            text_scale
-        )
-        self.text.next_to(
-            self.rect,
-            direction=ORIGIN,
-            coor_mask=text_position * 0.8,
-            aligned_edge=text_position,
-        )
-
-        super().__init__(self.rect, self.text, **kwargs)
-        # super().arrange(ORIGIN)
-
-
-class Pixel(Square):
-    def __init__(self, n: int, color_mode: str):
-        assert color_mode in ("RGB", "GRAY"), "Color modes are RGB and GRAY"
-
-        if color_mode == "RGB":
-            color = rgb_to_hex(n / 255)
-        else:
-            color = g2h(n / 255)
-        super().__init__(side_length=1)
-
-        self.set_stroke(BLACK, width=0.2)
-        self.set_fill(color, opacity=1)
-
-
-class PixelArray(VGroup):
-    def __init__(self, img: np.ndarray, include_numbers=False, color_mode="RGB"):
-
-        if len(img.shape) == 3:
-            rows, cols, channels = img.shape
-        else:
-            rows, cols = img.shape
-
-        self.shape = img.shape
-
-        pixels = []
-        for row in img:
-            for p in row:
-                if include_numbers:
-                    self.number = (
-                        Text(str(p), font="SF Mono")
-                        .scale(0.7)
-                        .set_color(g2h(1) if p < 180 else g2h(0))
-                    )
-                    pixels.append(VGroup(Pixel(p, color_mode), self.number))
-                else:
-                    pixels.append(Pixel(p, color_mode))
-
-        super().__init__(*pixels)
-        self.arrange_in_grid(rows, cols, buff=0)
-
-        self.dict = {index: p for index, p in enumerate(self)}
-
-    def __getitem__(self, value) -> VGroup:
-        if isinstance(value, slice):
-            return VGroup(*list(self.dict.values())[value])
-        else:
-            return self.dict[value]
 
 
 class IntroduceRGBAndJPEG(Scene):
@@ -2878,258 +2693,3 @@ class TestHexToGrayScale(Scene):
             self.add(dot)
             self.wait()
             self.remove(dot)
-
-
-def gray_scale_value_to_hex(value):
-    hex_string = hex(value).split("x")[-1]
-    if value < 16:
-        hex_string = "0" + hex_string
-    return "#" + hex_string * 3
-
-
-def make_lut_u():
-    return np.array([[[i, 255 - i, 0] for i in range(256)]], dtype=np.uint8)
-
-
-def make_lut_v():
-    return np.array([[[0, 255 - i, i] for i in range(256)]], dtype=np.uint8)
-
-
-def dct1D_manual(f, N):
-    result = []
-    constant = (2 / N) ** 0.5
-    for u in range(N):
-        component = 0
-        if u == 0:
-            factor = 1 / np.sqrt(2)
-        else:
-            factor = 1
-        for i in range(N):
-            component += (
-                constant * factor * np.cos(np.pi * u / (2 * N) * (2 * i + 1)) * f(i)
-            )
-
-        result.append(component)
-
-    return result
-
-
-def f(i):
-    return np.cos((2 * i + 1) * 3 * np.pi / 16)
-
-
-def plot_function(f, N):
-    import matplotlib.pyplot as plt
-
-    x = np.arange(0, N, 0.001)
-    y = f(x)
-
-    plt.plot(x, y)
-    plt.show()
-
-
-def g(i):
-    return np.cos((2 * i + 1) * 5 * np.pi / 16)
-
-
-def h(i):
-    return np.cos((2 * i + 1) * 3 * np.pi / 16) * np.cos((2 * i + 1) * 1.5 * np.pi / 16)
-
-
-def get_dct_elem(i, j):
-    return np.cos(j * (2 * i + 1) * np.pi / 16)
-
-
-def get_dct_matrix():
-    matrix = np.zeros((8, 8))
-    for i in range(8):
-        for j in range(8):
-            matrix[j][i] = get_dct_elem(i, j)
-
-    return matrix
-
-
-def get_dot_product_matrix():
-    dct_matrix = get_dct_matrix()
-    dot_product_matrix = np.zeros((8, 8))
-    for i in range(8):
-        for j in range(8):
-            value = np.dot(dct_matrix[i], dct_matrix[j])
-            if np.isclose(value, 0):
-                dot_product_matrix[i][j] = 0
-            else:
-                dot_product_matrix[i][j] = value
-
-    return dot_product_matrix
-
-
-def format_block(block):
-    if len(block.shape) < 3:
-        return block.astype(float) - 128
-    # [0, 255] -> [-128, 127]
-    block_centered = block[:, :, 1].astype(float) - 128
-    return block_centered
-
-
-def invert_format_block(block):
-    # [-128, 127] -> [0, 255]
-    new_block = block + 128
-    # in process of dct and inverse dct with quantization,
-    # some values can go out of range
-    new_block[new_block > 255] = 255
-    new_block[new_block < 0] = 0
-    return new_block
-
-
-def dct_1d(row):
-    return fftpack.dct(row, norm="ortho")
-
-
-def idct_1d(row):
-    return fftpack.idct(row, norm="ortho")
-
-
-def dct_2d(block):
-    return fftpack.dct(fftpack.dct(block.T, norm="ortho").T, norm="ortho")
-
-
-def idct_2d(block):
-    return fftpack.idct(fftpack.idct(block.T, norm="ortho").T, norm="ortho")
-
-
-def quantize(block):
-    quant_table = get_quantization_table()
-    return (block / quant_table).round().astype(np.int32)
-
-
-def get_quantization_table():
-    quant_table = np.array(
-        [
-            [16, 11, 10, 16, 24, 40, 51, 61],
-            [12, 12, 14, 19, 26, 58, 60, 55],
-            [14, 13, 16, 24, 40, 57, 69, 56],
-            [14, 17, 22, 29, 51, 87, 80, 62],
-            [18, 22, 37, 56, 68, 109, 103, 77],
-            [24, 35, 55, 64, 81, 104, 113, 92],
-            [49, 64, 78, 87, 103, 121, 120, 101],
-            [72, 92, 95, 98, 112, 100, 103, 99],
-        ]
-    )
-    return quant_table
-
-
-def dequantize(block):
-    quant_table = get_quantization_table()
-    return (block * quant_table).astype(np.float)
-
-
-def rgb2ycbcr(r, g, b):  # in (0,255) range
-    y = 0.299 * r + 0.587 * g + 0.114 * b
-    cb = 128 - 0.168736 * r - 0.331364 * g + 0.5 * b
-    cr = 128 + 0.5 * r - 0.418688 * g - 0.081312 * b
-    return y, cb, cr
-
-
-def ycbcr2rgb(y, cb, cr):
-    r = y + 1.402 * (cr - 128)
-    g = y - 0.34414 * (cb - 128) - 0.71414 * (cr - 128)
-    b = y + 1.772 * (cb - 128)
-    return r, g, b
-
-
-def g2h(n):
-    """Abbreviation for grayscale to hex"""
-    return rgb_to_hex((n, n, n))
-
-
-def coords2rgbcolor(i, j, k):
-    """
-    Function to transform coordinates in 3D space to hexadecimal RGB color.
-
-    @param: i - x coordinate
-    @param: j - y coordinate
-    @param: k - z coordinate
-    @return: hex value for the corresponding color
-    """
-    return rgb_to_hex(
-        (
-            (i) / 255,
-            (j) / 255,
-            (k) / 255,
-        )
-    )
-
-
-def coords2ycbcrcolor(i, j, k):
-    """
-    Function to transform coordinates in 3D space to hexadecimal YCbCr color.
-
-    @param: i - x coordinate
-    @param: j - y coordinate
-    @param: k - z coordinate
-    @return: hex value for the corresponding color
-    """
-    y, cb, cr = rgb2ycbcr(i, j, k)
-
-    return rgb_to_hex(
-        (
-            (y) / 255,
-            (cb) / 255,
-            (cr) / 255,
-        )
-    )
-
-
-def index2coords(n, base):
-    """
-    Changes the base of `n` to `base`, assuming n is input in base 10.
-    The result is then returned as coordinates in `len(result)` dimensions.
-
-    This function allows us to iterate over the color cubes sequentially, using
-    enumerate to index every cube, and convert the index of the cube to its corresponding
-    coordinate in space.
-
-    Example: if our ``color_res = 4``:
-
-        - Cube #0 is located at (0, 0, 0)
-        - Cube #15 is located at (0, 3, 3)
-        - Cube #53 is located at (3, 1, 1)
-
-    So, we input our index, and obtain coordinates.
-
-    @param: n - number to be converted
-    @param: base - base to change the input
-    @return: list - coordinates that the number represent in their corresponding space
-    """
-    if base == 10:
-        return n
-
-    result = 0
-    counter = 0
-
-    while n:
-        r = n % base
-        n //= base
-        result += r * 10 ** counter
-        counter += 1
-
-    coords = list(f"{result:03}")
-    return coords
-
-
-def get_zigzag_order(block_size=8):
-    return zigzag(block_size)
-
-
-def zigzag(n):
-    """zigzag rows"""
-
-    def compare(xy):
-        x, y = xy
-        return (x + y, -y if (x + y) % 2 else y)
-
-    xs = range(n)
-    return {
-        n: index
-        for n, index in enumerate(sorted(((x, y) for x in xs for y in xs), key=compare))
-    }
